drm/amd/display: Fix implicit conversion warning
[Why & How]
Fix implicit narrowing conversion warnings.

Tested-by: Daniel Wheeler <daniel.wheeler@amd.com>

+++ drivers/gpu/drm/amd/display/dc/core/dc_hw_sequencer.c
@@ -3531,7 +3531,7 @@ void hwss_send_cursor_info_to_dmu(union block_sequence_params *params)
 	struct pipe_ctx *pipe_ctx = params->send_cursor_info_to_dmu_params.pipe_ctx;
 	int pipe_idx = params->send_cursor_info_to_dmu_params.pipe_idx;
 
-	dc_send_update_cursor_info_to_dmu(pipe_ctx, pipe_idx);
+	dc_send_update_cursor_info_to_dmu(pipe_ctx, (uint8_t)pipe_idx);
 }
 
 void hwss_set_cursor_attribute(union block_sequence_params *params)
